requirements.txt
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.4.0
Pygments==2.19.2
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
PyYAML==6.0.3
RapidFuzz==3.14.1
rich==14.1.0
shiboken6==6.9.2
textual==6.1.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
